refactor: changed the filename of templateParameters file and its references
