+++ types/standard/config/webpack.config.js
@@ -17,7 +17,7 @@ const publicPath = '';
 const output = '';
 
 const projectConfig = require('./index.js');
-const defaultProperties = require('./constants')
+const templateParameters = require('./templateParameters')
 
 const isProject = !fs.existsSync(path.join(appDirectory, 'web', 'client', 'product'));
 
@@ -79,7 +79,7 @@ module.exports = () => {
                     filename: key,
                     template: projectConfig.htmlTemplates[key],
                     templateParameters: {
-                        ...defaultProperties,
+                        ...templateParameters,
                         ...projectConfig.templateParameters,
                         version: 'dev',
                         name: projectConfig.name
